extract ErrorMessage and Input from FomInput
