+++ src/components/Input/Input.module.scss
@@ -1,0 +1,52 @@
+.input_wrapper{
+  .label {
+    display: block;
+    font-size: 16px;
+    line-height: 24px;
+    letter-spacing: 0.15px;
+    margin-bottom: 8px;
+    text-align: left;
+  }
+
+  .input {
+    display: block;
+    width: 100%;
+    background-color: #ffffff;
+    color: rgb(41, 43, 50);
+    font-size: 14px;
+    line-height: 20px;
+    letter-spacing: 0.25px;
+    height: 42px;
+    padding: 4px 8px;
+    border-radius: 4px;
+    border: 1px solid #bbc2dc;
+    box-sizing: border-box;
+    outline: none;
+
+    &:hover,
+    &:focus,
+    &:focus-within {
+      border: 1px solid #3D8EB0;
+      box-shadow: 0 0 3px 2px rgba(1,164,209,.29);
+    }
+
+    &:-webkit-autofill,
+    &:-webkit-autofill:hover,
+    &:-webkit-autofill:focus,
+    &:-webkit-autofill:active {
+      -webkit-transition-delay: 99999s;
+      -webkit-transition: color 99999s ease-out, background-color 99999s ease-out;    
+    }
+  }
+
+  .error_input {
+    border-radius: 4px;
+    border: 1px solid #CE0E2D;
+
+    &:hover,
+    &:focus,
+    &:focus-within {
+      border: 1px solid #CE0E2D;
+    }
+  }
+}
