+++ src/pages/BrowseTask/component/CreateProfile/SubPages/Birthday/Birthday.module.scss
@@ -3,7 +3,7 @@
   flex-direction: column;
 
   .input_wrapper{
-    width: 60px;
+    width: 200px;
     padding: 0 4px;
     display: flex;
     margin: 0;
